requirements.txt
cvzone
numpy
opencv-python
pillow
tensorflow
keras